Release XPHearts v1.3.3 — withdraw command, +0.5x charm, hostile-only charging
Add /withdraw: players with >= 2.0x multiplier can extract exactly 1.0x as a
tradeable Multiplier Token (NETHER_STAR + NBT). Inventory-full rolls back.
Token is applied by holding it in offhand and right-clicking (+1.0x).

Change charm consume from +1.0x to +0.5x. Multiplier storage changed from
int to double throughout (PlayerDataManager, CharmListener, XPMultiplierCommand,
XPHeartsExpansion). Existing integer values in data.yml load cleanly.

Add allow-passive-mobs config (default: false). When false, only hostile mobs
(Monster interface) charge the charm. XP multiplier boost still applies to all
kills regardless of this setting.

+++ src/main/java/com/xphearts/data/PlayerDataManager.java
@@ -14,7 +14,7 @@ public class PlayerDataManager {
     private final XPHearts plugin;
     private final File dataFile;
     private YamlConfiguration data;
-    private final Map<UUID, Integer> cache = new HashMap<>();
+    private final Map<UUID, Double> cache = new HashMap<>();
 
     public PlayerDataManager(XPHearts plugin) {
         this.plugin = plugin;
@@ -34,7 +34,8 @@ public class PlayerDataManager {
         data = YamlConfiguration.loadConfiguration(dataFile);
         for (String key : data.getKeys(false)) {
             try {
-                cache.put(UUID.fromString(key), data.getInt(key + ".multiplier", 1));
+                // getDouble handles both legacy int values and new double values
+                cache.put(UUID.fromString(key), data.getDouble(key + ".multiplier", 1.0));
             } catch (IllegalArgumentException ignored) {}
         }
     }
@@ -48,18 +49,18 @@ public class PlayerDataManager {
         }
     }
 
-    public int getMultiplier(UUID uuid) {
-        return cache.getOrDefault(uuid, 1);
+    public double getMultiplier(UUID uuid) {
+        return cache.getOrDefault(uuid, 1.0);
     }
 
-    public void setMultiplier(UUID uuid, int value) {
-        int max = plugin.getConfig().getInt("multiplier.max-multiplier", 10);
-        cache.put(uuid, Math.max(1, Math.min(value, max)));
+    public void setMultiplier(UUID uuid, double value) {
+        double max = plugin.getConfig().getDouble("multiplier.max-multiplier", 10.0);
+        cache.put(uuid, Math.max(1.0, Math.min(value, max)));
         save();
     }
 
     public void resetMultiplier(UUID uuid) {
-        cache.put(uuid, 1);
+        cache.put(uuid, 1.0);
         save();
     }
 }
